+++ gems.rb
@@ -9,7 +9,7 @@ gem "pg",                   "0.21.0"
 
 # Gems used by project
 gem "autoprefixer-rails"
-gem "bootstrap",            "~> 4.0.0.beta"
+gem "bootstrap",            "~> 4.0.0.beta2"
 gem "carrierwave",          "~> 1.2.1"
 gem "chunky_png",           "~> 1.3.4"
 gem "devise",               "~> 4.3.0"