+++ CHANGELOG.md
@@ -82,7 +82,7 @@
 - **Gem Changes**
   - Updated to ruby 2.4.2
   - Updated to rails 5.1.4
-  - Updated to bootstrap 4.0.0.beta
+  - Updated to bootstrap 4.0.0.beta2
   - Updated to carrierwave 1.2.1
   - Updated to haml 5.0.4
   - Updated to kaminari 1.1.0